+++ src/Register.zig
@@ -3,7 +3,7 @@ const std = @import("std");
 name: []const u8,
 description: ?[]const u8 = null,
 addr_offset: u64,
-size: ?usize = null,
+size: ?u64 = null,
 access: ?@import("svd.zig").Access = null,
 reset_value: ?u64 = null,
 reset_mask: ?u64 = null,
